CHROMIUM: kbuild: add forward declaration of default target to Makefile.headersinst
In make the "default" target is the one that is declared first.  Prior
to the commit ("kbuild: shrink .cache.mk when it exceeds 1000 lines"),
including Kbuild.include didn't declare any rules.  Thus the first
rule declared in Makefile.headersinst could have been "__headersinst"
or "__headerscheck".  After the aforementioned commit the first
declared rule will be one saying that make shouldn't try to generate
the ".cache.mk" file.

In order to go back to the way things used to be, we just need to add
a predeclaration of the right rule.  This is basically the same thing
as in the upstream commit fa9acf703d64 ("kbuild: add forward
declaration of default target to Makefile.asm-generic").

NOTE: this patch isn't needed upstream because upstream declares a
target at the top of Makefile.headersinst.  ...thus, this is a
CHROMIUM patch.  If this ever ends up causing conflicts then
presumably you're picking a bunch of upstream patches so you'll end up
with a proper early declaration.  Thus, you should just be able to
revert this.  :)

BUG=None
TEST='git clean -d -x -f -q; make defconfig; make; make allyesconfig; make'

Change-Id: I1ee497b680299b73585c984f1c7828c5c757c0cc
Signed-off-by: Douglas Anderson <dianders@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/738836
Reviewed-by: Guenter Roeck <groeck@chromium.org>
Signed-off-by: Nathan Chancellor <natechancellor@gmail.com>